fix(server): use project data for document links
Resolve symbol and kwarg document links from project AST, type, and enum
information. Migrate general document link tests to plain XGo and the
isolated classfile framework without requiring embedded package data.

Keep spx resource links and their regressions in dedicated files. Use
classfile registrations to distinguish spx from another framework using
`.spx`, preserving resource filtering and link metadata.

Cover source file kinds, cross-file fields and enums, framework calls,
missing documentation, synthetic symbols, document updates, incomplete
source, and UTF-16 ranges.

Wait for background diagnostics in the notification ordering test so it
cannot race with later tests that use shared imported packages.
